requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
Django==4.1.7
django-ckeditor==6.5.1
django-constance==2.9.1
django-extensions==3.2.1
django-js-asset==2.0.0
django-picklefield==3.1
Pillow==9.4.0
redis==4.5.1
sqlparse==0.4.3
